@@ -21,12 +21,12 @@
     ],
     "homepage": "https://www.mediatis.de",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "digital-marketing-framework/core": "dev-feature/MFR-136-t3-13",
         "digital-marketing-framework/template-engine-twig": "dev-feature/MFR-136-t3-13"
     },
     "require-dev": {
-        "mediatis/coding-standards": "^1.0"
+        "mediatis/coding-standards": "^2.0"
     },
     "autoload": {
         "psr-4": {
